refactor: extract writeExclusion/hasExclusion to shared src/contacts/dedup-exclusions module
Moves the exclusion helpers out of the one-off sweep script into a shared module
so the contacts agent's skill can import the same implementation without duplication.
The script re-exports them for backward compatibility. Adds optional source param
(default 'contacts-dedup') so skills can pass ctx.memoryWriteSource for correct
per-task rate-limit tracking.
